Update E2E tests
This commit introduces a bigger separation between the PO and the E2E tests than was present in the previous version. The E2E tests do not need to know how the page is actually structured, so rather than typing in an input field, for example, this new E2E test says to filter by age 27 (which is currently typed in an input field, but could be from a dropdown, for example).